* 'feature/Q&A' of https://github.com/zhiyicx/thinksns-plus-android:
  (#674) 添加分享功能，修改话题标签
  (#678)    code		update reward amount type

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/remote/InfoMainClient.java
@@ -206,7 +206,7 @@ public interface InfoMainClient {
      */
     @FormUrlEncoded
     @POST(APP_PATH_INFO_REWARDS)
-    Observable<Object> rewardInfo(@Path("news_id") long news_id, @Field("amount") float amount);
+    Observable<Object> rewardInfo(@Path("news_id") long news_id, @Field("amount") Integer amount);
 
 
     /**
